Fixes breakpoint setting on Windows
After dap-mode commit 5e449c864107e08353fd21c44897c4d480190d94
dap-mode no longer replaces the `/` in filenames by `\\`, which is
needed for the java debug server to recognize the path as a
path (insted of as an URI).

Currently dap-mode sends the setBreakpoints request with filenames
like c:/users/myuser/app.java, which the java debug server believes to
be an URI of the c scheme and it fails to set the breakpoint.

This change will define a local-to-remote-path-fn in the default
dap--debug-session that will handle the conversion.